spawn timers using deltaTime and warnings removed

+++ src/entities/components/behavior_tree/actions/rotate_aim.h
@@ -11,6 +11,6 @@ protected:
 	virtual Status update(float step) override;
 
 private:
-	int mAngle;
+	float mAngle;
 };
